Improving the structure of the project
types folder was moved inside include, Utils was moved inside src and text related files split in Text folder, created Path file to create function so that the executable can be run from any folder. Assets folder now contains an assets folder with the PNG's and a font folder with the .ttf. The .o are now in bin-int folder, and its subdirectories, Makefile was updated to handle these changes